Create api module and move GetDBEndpoints to it
GetDBEndpoints will also be used by other operators like
neutron, octavia etc.
